File Division and Reorganization
Divide header files and reorganize them.
I'm going to try to classify the files that are constituted for only header files and header files and implementation files, because it is just convenient aftertime. So I plan to make a branch and do it.
In addition, I think that a default value of template parameter should be written at its forward declaration, because it probably reduces dependency between their header files. In the coding style of Boost, *Fwd.* files that have its forward declaration are placed, so I learn from it and reorganize my project.