Divide Memory Management Functions, Add a Utility Class
About the behavior of HeapProvider, I understood that an using memory management function other than SimpleHeap<T, DefaultHeap> made no sense, as the result of value the ability to the retrieval performance with any keys. So, I added the new class SimpleHeapProvider that was value the ability to the generative performance.