Investigate the Way of Generating Identifier for Delay Evaluation
Investigate the way of generating the objects that represent an identifier to use for delay evaluation.
I groped the action against the above thing because the objects comparatively live for a long time but they are small and generated in large numbers.
The result is in Test.Urasandesu.CppAnonym.cpp, but it is like a scratched code.
Finally, I plan to merge the way of using Boost.Pool because it indicates very good performance.