Reorganize Memory Management Strategy
With the growth of structural complexity, I understood that the way of heap memory management using boost::shared_ptr increases the risk to occur memory leak.
However, if boost::weak_ptr is used to reduce this risk, the readability of source code is decreased.
In the first place, I can detect that the disposed object has problem, because the objects have already been allocated by boost::pool. Therefore, I manage these objects with raw pointer.

For your information, the auto generation for the hierarchal structure represented by SmartPtrChain is kept, because it will reduce the unnecessary repeat.